[MOD/#11] 코드 개선
