+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.MainActivity"
+            android:name=".presentation.onboarding.Activity.OnBoardingActivity"
             android:exported="true"
             android:screenOrientation="portrait">
             <intent-filter>
@@ -27,7 +27,7 @@
             </intent-filter>
         </activity>
         <activity android:name=".presentation.yello.YelloActivity"
-            android:exported="false" />z
+            android:exported="false" />
     </application>
 
 </manifest>
