+++ app/src/main/java/com/yello/presentation/onboarding/fragment/GenderFragment.kt
@@ -1,19 +1,19 @@
-package com.yello.presentation.onboarding
-
+package com.yello.presentation.onboarding.fragment
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.ui.base.BindingFragment
 import com.yello.R
-import com.yello.databinding.FragmentStudentidBinding
+import com.yello.databinding.FragmentGenderBinding
 
-class StudentIdFragment : BindingFragment<FragmentStudentidBinding>(R.layout.fragment_studentid) {
+class GenderFragment : BindingFragment<FragmentGenderBinding>(R.layout.fragment_gender) {
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        return inflater.inflate(R.layout.fragment_studentid, container, false)
+        return inflater.inflate(R.layout.fragment_gender, container, false)
     }
 }
