+++ app/src/main/java/com/yello/presentation/onboarding/fragment/SchoolFragment.kt
@@ -1,4 +1,4 @@
-package com.yello.presentation.onboarding
+package com.yello.presentation.onboarding.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,14 +6,15 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.ui.base.BindingFragment
 import com.yello.R
-import com.yello.databinding.FragmentNameIdBinding
+import com.yello.databinding.FragmentSchoolBinding
 
-class NameIdFragment : BindingFragment<FragmentNameIdBinding>(R.layout.fragment_name_id) {
+class SchoolFragment : BindingFragment<FragmentSchoolBinding>(R.layout.fragment_school) {
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        return inflater.inflate(R.layout.fragment_name_id, container, false)
+        return inflater.inflate(R.layout.fragment_school, container, false)
     }
 }
