+++ app/src/main/java/com/yello/presentation/onboarding/fragment/StudentIdFragment.kt
@@ -1,4 +1,4 @@
-package com.yello.presentation.onboarding
+package com.yello.presentation.onboarding.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,15 +6,14 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.ui.base.BindingFragment
 import com.yello.R
-import com.yello.databinding.FragmentStartappBinding
+import com.yello.databinding.FragmentStudentidBinding
 
-class StartAppFragment : BindingFragment<FragmentStartappBinding>(R.layout.fragment_startapp) {
-
+class StudentIdFragment : BindingFragment<FragmentStudentidBinding>(R.layout.fragment_studentid) {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        return inflater.inflate(R.layout.fragment_startapp, container, false)
+        return inflater.inflate(R.layout.fragment_studentid, container, false)
     }
 }
