+++ app/src/main/res/layout/fragment_startapp.xml
@@ -26,7 +26,7 @@
             android:id="@+id/btn_startappnext"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="완"
+            android:text="완료"
             app:layout_constraintBottom_toBottomOf="parent" />
